+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.0.2] - rev 85
+## [0.0.2] - rev 88
 
+- alter.ini: new option CHARTS_NETDESC that enables networks description render on charts
 - installer: FreeBSD 13.2 setup works again.
 - installer: FreeBSD 13.3 setup tested and works
 - installer: Linux Debian 12.5 bookworm installer
 - settings: networks descriptions implemented
+- graph: optional ability to render network description or its CIDR on charts
 
 ## [0.0.1] - rev 77